+++ R/yPlot.R
@@ -72,10 +72,15 @@
 #' is to make the target data into a factor, and to put its levels in the desired order: \code{factor(data, levels = c("level1", "level2", ...))}.
 #' @param x.labels.rotate Logical which sets whether the labels should be rotated.
 #' Default: \code{TRUE} for violin and box plots, but \code{FALSE} for ridgeplots.
-#' @param add.line numeric value(s) where one or multiple line(s) should be added
+#' @param add.line Numeric value(s), denoting y-axis value(s), where one or multiple horizonal line(s) should be added.
 #' @param line.linetype String which sets the type of line for \code{add.line}.
 #' Defaults to "dashed", but any ggplot linetype will work.
-#' @param line.color String that sets the color(s) of the \code{add.line} line(s)
+#' @param line.color String that sets the color(s) of the \code{add.line} line(s). Default = "black".
+#' Alternatively, a vector of strings of the same length as \code{add.line} can be given to set the color of each line individually.
+#' @param line.linewidth Number that sets the thickness of the \code{add.line} line(s). Default = 0.5.
+#' Alternatively, a vector of numbers of the same length as \code{add.line} can be given to set the thickness of each line individually.
+#' @param line.opacity Number that sets the opacity of the \code{add.line} line(s). Default = 1.
+#' Alternatively, a vector of numbers of the same length as \code{add.line} can be given to set the opacity of each line individually.
 #' @param jitter.size Scalar which sets the size of the jitter shapes.
 #' @param jitter.width Scalar that sets the width/spread of the jitter in the x direction. Ignored in ridgeplots.
 #'
@@ -104,7 +109,7 @@
 #' @param vlnplot.width Scalar which sets the width/spread of violin plots in the x direction
 #' @param vlnplot.scaling String which sets how the widths of the of violin plots are set in relation to each other.
 #' Options are "area", "count", and "width". If the default is not right for your data, I recommend trying "width".
-#' For an explanation of each, see \code{\link{geom_violin}}.
+#' For an explanation of each, see \code{\link[ggplot2]{geom_violin}}.
 #' @param vlnplot.quantiles Single number or numeric vector of values in [0,1] naming quantiles at which to draw a horizontal line within each violin plot. Example: \code{c(0.1, 0.5, 0.9)}
 #' @param ridgeplot.lineweight Scalar which sets the thickness of the ridgeplot outline.
 #' @param ridgeplot.scale Scalar which sets the distance/overlap between ridgeplots.
@@ -259,6 +264,9 @@
 #'     main = "CD3E",
 #'     legend.show = FALSE)
 #'
+#' \dontrun{
+#' # (Due to unfortunate CRAN submission constraints)
+#'
 #' # Data can also be split in other ways with 'shape.by' or 'split.by'
 #' yPlot(data_frame = example_df, var = "gene1", group.by = "timepoint",
 #'     plots = c("vlnplot", "boxplot", "jitter"),
@@ -281,7 +289,9 @@
 #'     var = c("gene1", "gene2"),
 #'     multivar.aes = "color")
 #'
-#' @author Daniel Bunis
+#' }
+#'
+#' @author Daniel Bunis, Jared Andrews
 #' @export
 
 yPlot <- function(
@@ -348,6 +358,8 @@ yPlot <- function(
     add.line = NULL,
     line.linetype = "dashed",
     line.color = "black",
+    line.linewidth = 0.5,
+    line.opacity = 1,
     add.pvalues = NULL,
     pvalues.subgroup.method = c("color", "group", "ignore"),
     pvalues.round.digits = 4,
@@ -515,7 +527,7 @@ yPlot <- function(
             boxplot.outlier.size, boxplot.fill,
             boxplot.position.dodge, boxplot.lineweight,
             vlnplot.lineweight, vlnplot.width, vlnplot.scaling,
-            vlnplot.quantiles, add.line, line.linetype, line.color,
+            vlnplot.quantiles,
             x.labels.rotate, do.hover, y.breaks, min, max, data_frame,
             cols_use$group.aes,
             stats_plot, cols_use$p.by, pvalues.offset.above, pvalues.plot.adjust)
@@ -535,6 +547,19 @@ yPlot <- function(
         p <- .add_splitting(
             p, cols_use$split.by, split.nrow, split.ncol, split.adjust)
     }
+
+    # Get number of panels so that replicates of aesthetics can be generated if supplied for each line.
+    pp <- ggplot_build(p)
+    num.panels <- length(levels(pp$data[[1]]$PANEL))
+
+    if (!is.null(add.line)) {
+        if(!("ridgeplot" %in% plots)) {
+            p <- .add_yline(p, add.line, line.linetype, line.color, line.linewidth, line.opacity, num.panels)
+        } else {
+            p <- .add_xline(p, add.line, line.linetype, line.color, line.linewidth, line.opacity, num.panels)
+        }
+    }
+
     if (!legend.show) {
         p <- .remove_legend(p)
     }
@@ -567,7 +592,6 @@ yPlot <- function(
     boxplot.width, boxplot.color, boxplot.show.outliers, boxplot.outlier.size,
     boxplot.fill, boxplot.position.dodge, boxplot.lineweight,
     vlnplot.lineweight, vlnplot.width, vlnplot.scaling, vlnplot.quantiles,
-    add.line, line.linetype, line.color,
     x.labels.rotate, do.hover, y.breaks, min, max,
     data_frame, group.aes,
     stats = NULL, p.by, pvalues.offset.above, pvalues.plot.adjust
@@ -589,12 +613,21 @@ yPlot <- function(
     # Add Plots
     for (i in seq_along(plots)) {
         if (plots[i] == "vlnplot") {
-            p <- p + geom_violin(
+            violin.args <- list(
                 linewidth = vlnplot.lineweight,
                 width = vlnplot.width,
                 scale = vlnplot.scaling,
-                draw_quantiles = vlnplot.quantiles,
-                na.rm = TRUE)
+                na.rm = TRUE
+            )
+            if (!identical(vlnplot.quantiles, NULL)) {
+                .error_if_no_mass_because("quantiles to violins")
+                if ("quantile.linetype" %in% names(formals(ggplot2::geom_violin))) {
+                    violin.args$quantiles <- vlnplot.quantiles
+                } else {
+                    violin.args$draw_quantiles <- vlnplot.quantiles
+                }
+            }
+            p <- p + do.call(geom_violin, violin.args)
         }
 
         if (plots[i] == "boxplot") {
@@ -677,9 +710,6 @@ yPlot <- function(
     if (is.na(x.labels.rotate) || x.labels.rotate) {
         p <- p + theme(axis.text.x= element_text(angle=45, hjust = 1, vjust = 1))
     }
-    if (!is.null(add.line)) {
-        p <- p + geom_hline(yintercept=add.line, linetype= line.linetype, color = line.color)
-    }
 
     # Add p-value brackets
     if (!is.null(stats)) {
@@ -761,9 +791,6 @@ yPlot <- function(
     p <- p + xlab(ylab) + ylab(xlab)
     if (!is.na(x.labels.rotate) && x.labels.rotate) {
         p <- p + theme(axis.text.y= element_text(angle=45, hjust = 1, vjust = 1))
-    }
-    if (!is.null(add.line)) {
-        p <- p + geom_vline(xintercept=add.line, linetype= line.linetype, color = line.color)
     }
 
     if (!is.null(stats)) {