+++ NEWS.md
@@ -1,3 +1,18 @@
+# dittoViz 1.0.5
+
+(All relating to integrating dittoViz functions with dittoSeq)
+* Minor warning and documentation adjustments.
+* BugFix: ignore 'do.letter' when conflicts with 'shape.by' or 'do.hover' use
+* BugFix: Allow 'labels.use.numbers = TRUE' to work alongside of 'legend.color.title = NULL'
+
+# dittoViz 1.0.4
+
+* Minor changes to ensure compatibility with ggplot2 version 4.0.0 changes. Notably, the 'MASS' package is now added as a suggested companion package, and error messages suggesting its installation were added for features which rely on functionality from this package. This change is concurrent with ggplot2 itself downgrading the tool from an imported requirement to a suggested addition.
+
+# dittoViz 1.0.3
+
+* Expanded flexibility when adding lines to plots: Adds ability to place sloped lines with scatterPlot() and scatterHex() via new 'add.abline' and 'abline.*' inputs, as well as previously unavailable opacity and lineweight control inputs for all existing 'add.line' variants.
+
 # dittoViz 1.0.2
 
 * For scatterPlot() and scatterHex(), added the ability to use numbers as data labels instead of the data-values themselves, because sometimes data-values are quite long. Controlled with new inputs 'labels.use.numbers' and 'labels.numbers.spacer'.